+++ Extension/Xamarin.Forms.HotReload.Extension.Rider/CHANGELOG.md
@@ -1,2 +1,6 @@
 # Changelog
-First version of HotReload plugin for JetBrains Rider IDE.
+
+## 1.3.0
+- Added notification message about HotReload observer port
+- Plugin work was fixed in case when multiple instances of IDE are open
+- Plugin version was synced with other extensions